requirements.txt
alembic==1.13.1
async-timeout==4.0.3
asyncpg==0.29.0
bash==0.6
blinker==1.8.2
botocore==1.34.115
cached-property==1.5.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorclass==2.2.2
configparser==7.0.0
Deprecated==1.2.14
dnspython==2.6.1
docopt==0.6.2
flash==1.0.3
flask==3.0.3
Flask-Email==1.4.4
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.7.0
Flask-Login==0.6.3
flask-mail==0.10.0
Flask-Migrate==2.5.2
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.4.1
flask-wtf==1.2.1
greenlet==3.0.3
gunicorn==22.0.0
idna==3.7
importlib-metadata==7.1.0
importlib-resources==6.4.0
inflection==0.5.1
itsdangerous==2.2.0
jinja2==3.1.4
jmespath==1.0.1
limits==3.12.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
mypy-extensions==1.0.0
ordered-set==4.1.0
packaging==24.0
pillow==10.3.0
pip-upgrader==1.4.15
psycopg2==2.8.4
psycopg2-binary==2.9.9
pygments==2.18.0
PyJWT==2.8.0
pymongo==4.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
razorpay==1.4.2
redmail==0.6.0
requests==2.32.3
rich==13.7.1
s3transfer==0.10.1
six==1.16.0
SQLAlchemy==2.0.30
sqlalchemy-orm==1.2.10
SQLAlchemy-Utils==0.41.2
terminaltables==3.1.10
typing-extensions==4.12.0
typing-inspect==0.9.0
urllib3==1.26.18
uuid==1.30
werkzeug==3.0.3
wrapt==1.16.0
wtforms==3.1.2
zipp==3.19.0